Update Constant folding to new fixture testing
