+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/basic/actual.js
@@ -1,0 +1,8 @@
+"a" + "b";
+2 * 3;
+1 / 3;
+4 | 3;
+a(), b();
+var x = 1;
+foo(x);
+"b" + a + "c" + "d" + g + z + "f" + "h" + "z";
