+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/basic/expected.js
@@ -1,0 +1,8 @@
+"ab";
+6;
+1 / 3;
+7;
+a(), b();
+var x = 1;
+foo(x);
+"b" + a + "cd" + g + z + "fhz";
