+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/html-comment-escape/expected.js
@@ -1,0 +1,1 @@
+"<!--";
