+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/issue-440/expected.js
@@ -1,0 +1,1 @@
+var x = "'cool'test";
