+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/negative-zero/actual.js
@@ -1,0 +1,3 @@
+-0;
++-0;
++0;
