+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/runtime-errors/actual.js
@@ -1,0 +1,7 @@
+try {
+  x(
+    {
+      toString: 0
+    } + ""
+  );
+} catch (e) {}
