+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/runtime-errors/expected.js
@@ -1,0 +1,5 @@
+try {
+  x({
+    toString: 0
+  } + "");
+} catch (e) {}
