+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/script-escape/expected.js
@@ -1,0 +1,1 @@
+"</script";
