+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/spread-element/expected.js
@@ -1,0 +1,9 @@
+function foo() {
+  return [...iter].length;
+}
+function bar() {
+  return [...iter][0];
+}
+
+[...bar].reverse();
+[...foo].join("a");
