+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/string-literal-methods/actual.js
@@ -1,0 +1,16 @@
+"a,b,c".split(",");
+"a,b,c".split("");
+"a,b,c".split();
+"abc"[0];
+"abc"["0"];
+"abc"[4];
+"abc".charAt();
+"abc".charAt(1);
+"abc".charCodeAt();
+"abc".charCodeAt(1);
+"abc".length;
+
+"\u{1f44d}".charCodeAt();
+"\u{1f44d}".charCodeAt(1);
+"\u{1f44d}".codePointAt();
+"\u{1f44d}".codePointAt(1);
