+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/string-literal-methods/expected.js
@@ -1,0 +1,16 @@
+["a", "b", "c"];
+["a", ",", "b", ",", "c"];
+["a,b,c"];
+"a";
+"a";
+void 0;
+"a";
+"b";
+97;
+98;
+3;
+
+55357;
+56397;
+128077;
+56397;
