+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/style-escape/actual.js
@@ -1,0 +1,1 @@
+"</" + "style";
