+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/style-escape/expected.js
@@ -1,0 +1,1 @@
+"</style";
