+++ packages/babel-plugin-minify-constant-folding/__tests__/fixtures/to-string/expected.js
@@ -1,0 +1,6 @@
+"foo".toString();
+["foo", "bar"].toString();
+({}).toString();
+"foo".toString;
+["foo", "bar"].toString;
+({}).toString;
