+++ packages/babel-plugin-minify-constant-folding/__tests__/index.js
@@ -1,0 +1,2 @@
+const runner = require("test-runner");
+runner(__dirname);
